# riddler/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Edium Riddler
  version: "1.0.0"
servers:
  - url: /riddler/v1
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: QUIZ_NOT_FOUND
        description:
          type: string
          example: Квиз не найден
        details:
          type: object
          additionalProperties: true
...
    AnswerOptionForStudent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
...
    QuestionForStudent:
      type: object
      description: |
        Вопрос для студента — без правильных ответов.
        - single_choice / multiple_choice: `options` без `is_correct`
        - drag: `metadata.items` — перемешанные элементы
        - connection: `metadata.left` + `metadata.right` (правая колонка перемешана)
        - with_given_answer / with_free_answer: нет `options`, нет `metadata`
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [single_choice, multiple_choice, with_given_answer, with_free_answer, drag, connection]
        text:
          type: string
        image_id:
          type: string
          format: uuid
          nullable: true
          description: UUID изображения из сервиса Louvre
        max_score:
          type: integer
        options:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/AnswerOptionForStudent"
        metadata:
          type: object
          nullable: true
    AttemptSummary:
      type: object
      properties:
        attempt_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Нет в JSON, если в БД у попытки user_id NULL.
        status:
          type: string
          enum: [in_progress, grading, graded, completed, published]
        score:
          type: number
          format: float
          nullable: true
    AnswerReview:
      type: object
      properties:
        submission_id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        question_type:
          type: string
          enum: [single_choice, multiple_choice, with_given_answer, with_free_answer, drag, connection]
        question_text:
          type: string
        answer_data:
          type: object
          description: Ответ студента, формат зависит от типа вопроса
        final_score:
          type: number
          format: float
          nullable: true
        final_source:
          type: string
          nullable: true
          enum: [auto, llm, teacher]
        final_feedback:
          type: string
          nullable: true
        options:
          type: array
          description: Только для учителя или студента после завершенной проверки — варианты ответа с правильными отметками
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              text:
                type: string
              is_correct:
                type: boolean
        student_options:
          type: array
          description: Для студента — варианты ответа без правильных отметок
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              text:
                type: string
        metadata:
          type: object
          nullable: true
          description: Только для учителя или студента после завершенной проверки — correct_answers / correct_order / correct_pairs
    AttemptReview:
      type: object
      properties:
        attempt_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Нет в JSON, если в БД у попытки user_id NULL.
        status:
          type: string
          enum: [in_progress, grading, graded, completed, published]
        score:
          type: number
          format: float
          nullable: true
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerReview"
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Конфликт состояния
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /sessions/dashboard:
    get:
      tags: [Sessions]
      summary: Дашборд студента
      operationId: getStudentDashboard
      description: |
        Возвращает два списка для главного экрана студента:
        - `recent_grades` — последние 5 завершённых тестов (source=course), отсортированы по дате финиша
        - `active_tests` — доступные тесты (source=course, не истёк дедлайн), сначала те где есть попытка `in_progress`
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Дашборд
          content:
            application/json:
              schema:
                type: object
                properties:
                  recent_grades:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: string
                          format: uuid
                        quiz_template_id:
                          type: string
                          format: uuid
                        quiz_title:
                          type: string
                        attempt_id:
                          type: string
                          format: uuid
                        score:
                          type: number
                          nullable: true
                          description: Нормированная оценка 0–10
                        status:
                          type: string
                          enum: [grading, graded, completed, published]
                        finished_at:
                          type: string
                          format: date-time
                          nullable: true
                  active_tests:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: string
                          format: uuid
                        quiz_template_id:
                          type: string
                          format: uuid
                        quiz_title:
                          type: string
                        total_time_limit_sec:
                          type: integer
                          nullable: true
                        session_started_at:
                          type: string
                          format: date-time
                          nullable: true
                        session_finished_at:
                          type: string
                          format: date-time
                          nullable: true
                          description: Дедлайн сессии
                        attempt_id:
                          type: string
                          format: uuid
                          nullable: true
                        attempt_status:
                          type: string
                          nullable: true
                          enum: [in_progress]
        "401": {$ref: "#/components/responses/Unauthorized"}
  /sessions/awaiting-review:
    get:
      tags: [Sessions]
      summary: Сессии, ожидающие проверки
      operationId: listAwaitingReview
      description: |
        Возвращает тестовые сессии, созданные авторизованным пользователем,
        у которых есть хотя бы одна попытка в статусе `grading` или `graded`.
        Включает счётчики попыток по статусам для отображения прогресса проверки.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список сессий
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: string
                          format: uuid
                        quiz_template_id:
                          type: string
                          format: uuid
                        quiz_title:
                          type: string
                        grading_count:
                          type: integer
                          description: Попытки в процессе оценивания (LLM)
                        graded_count:
                          type: integer
                          description: Попытки, оценённые LLM, ожидают финальной проверки учителя
                        completed_count:
                          type: integer
                          description: Попытки, полностью проверенные учителем
        "401": {$ref: "#/components/responses/Unauthorized"}
...
  /sessions/{session_id}/attempts:
    post:
      tags: [Attempts]
      summary: Начать попытку
      operationId: createAttempt
      description: |
        Создаёт попытку и возвращает вопросы без правильных ответов.
        Порядок вопросов фиксируется на момент создания.
        Сессия должна быть в статусе `active`.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Попытка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt_id:
                    type: string
                    format: uuid
                  questions:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestionForStudent"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /attempts/{attempt_id}/answers:
    post:
      tags: [Attempts]
      summary: Отправить ответ
      operationId: submitAnswer
      description: |
        Upsert — повторный вызов с тем же `question_id` перезаписывает ответ.
        При истёкшем таймере попытка автозавершается, возвращает `ATTEMPT_EXPIRED`.
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question_id, answer_data]
              properties:
                question_id:
                  type: string
                  format: uuid
                answer_data:
                  type: object
                  description: |
                    Зависит от типа вопроса:
                    - single_choice: `{"selected_option_id": "uuid"}`
                    - multiple_choice: `{"selected_option_ids": ["uuid1", "uuid2"]}`
                    - with_given_answer / with_free_answer: `{"text": "ответ"}`
                    - drag: `{"order": ["B", "A", "C"]}`
                    - connection: `{"pairs": {"A": "1", "B": "2"}}`
      responses:
        "200":
          description: Ответ сохранён
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /attempts/{attempt_id}/finish:
    post:
      tags: [Attempts]
      summary: Завершить попытку
      operationId: finishAttempt
      description: |
        Завершает попытку и запускает оценку.

        Поведение зависит от настройки квиза `need_evaluation`:
        - **false** — автооценка сразу, попытка переходит в `completed`. `with_free_answer` не оценивается автоматически.
        - **true** — автооценка вопросов с вариантами ответов, попытка переходит в `grading`. Свободные ответы уходят в Charon (LLM). После завершения всех LLM-оценок попытка переходит в `graded`. Учитель затем проверяет и завершает (`completed`).
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Попытка завершена
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
    get:
      tags: [Teachers]
      summary: Список попыток сессии
      operationId: listSessionAttempts
      description: Только для учителя — автора квиза, к которому привязана сессия.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список попыток
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttemptSummary"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /attempts/{attempt_id}/review:
    get:
      tags: [Teachers]
      summary: Детали попытки с ответами и оценками
      operationId: getAttemptReview
      description: |
        Универсальный endpoint просмотра попытки.

        **Студент** (владелец попытки) — доступен при `status != in_progress`.
        Возвращает свои ответы, баллы и фидбек.
        - При `status = completed` — возвращает `options` с `is_correct` и `metadata` с правильными ответами.
        - При других статусах — возвращает `student_options` без `is_correct`, без `metadata`.

        **Учитель** (автор квиза) — доступен в любой момент.
        Дополнительно возвращает `options` с `is_correct` и `metadata` с правильными ответами.

        Без авторизации — только если у попытки в БД **`user_id` IS NULL** (тот же студенческий ответ, с той же логикой `student_options` или `options`/`metadata` в зависимости от статуса;
        только при `status != in_progress`). Иначе нужен JWT.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Детали попытки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptReview"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /attempts/{attempt_id}/grade:
    post:
      tags: [Teachers]
      summary: Выставить оценки за ответы попытки
      operationId: gradeAttempt
      description: |
        Учитель выставляет оценки за список ответов студента одним запросом.
        Перезаписывает auto/llm-оценку. Выполняется атомарно — при ошибке откатываются все изменения.
        Не отправляет итог в Caesar — для этого нужен `/complete`.
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grades]
              properties:
                grades:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [submission_id, score]
                    properties:
                      submission_id:
                        type: string
                        format: uuid
                      score:
                        type: number
                        format: float
                        minimum: 0
                        description: Балл от 0 до max_score вопроса
                      feedback:
                        type: string
                        nullable: true
                        description: Текстовый комментарий к оценке
      responses:
        "200":
          description: Оценки сохранены
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
  /attempts/session/{session_id}/publish:
    post:
      tags: [Teachers]
      summary: Опубликовать оценки по сессии
      operationId: publishSession
      description: |
        Переводит все `completed` попытки сессии в `published`.
        Требует: все не-кикнутые попытки в статусе `completed`.
        Отправляет событие `riddler.attempt.scored` для каждой попытки → Caesar обновляет оценки.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Оценки опубликованы
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...